+++ packages/apipgen-parser-swagger/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @apipgen/swag-parser
 
+## 0.1.5
+
+### Patch Changes
+
+- fix: filed name syntax parse error
+- Updated dependencies
+  - apipgen@0.3.7
+
 ## 0.1.4
 
 ### Patch Changes