+++ packages/apipgen-pipeline/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @apipgen/pipeline
 
+## 0.1.4
+
+### Patch Changes
+
+- fix: filed name syntax parse error
+- Updated dependencies
+  - apipgen@0.3.7
+
 ## 0.1.3
 
 ### Patch Changes