+++ packages/swag-fetch-ts/CHANGELOG.md
@@ -1,5 +1,14 @@
 # apipgen-swag-fetch-ts
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @apipgen/swag-parser@0.1.5
+  - @apipgen/pipeline@0.1.4
+  - apipgen@0.3.7
+
 ## 0.1.5
 
 ### Patch Changes