+++ packages/swag-got-ts/CHANGELOG.md
@@ -1,5 +1,14 @@
 # apipgen-swag-got-ts
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @apipgen/swag-parser@0.1.5
+  - @apipgen/pipeline@0.1.4
+  - apipgen@0.3.7
+
 ## 0.0.1
 
 ### Patch Changes